Scaffold Field Getters - Change: If the fields are being obtained in the context of a form binding allow  this binding is passed to the scaffold getter. Update controller scaffolding to pass the record.
Note: this changes the declaration order for scaffold fields, so the field params do not exist by default in the filter chain for those actions. While this is a bit of a PITA in some contexts, it does force more of the logic model side, so for the greater good really.